+++ application/views/tema/rodape.php
@@ -8,6 +8,14 @@
   <script src="<?= base_url(); ?>assets/js/bootstrap.min.js"></script>
   <script src="<?= base_url(); ?>assets/js/matrix.js"></script>
 </body>
-
+<script type="text/javascript">
+    $(document).ready(function() {
+        $('#tabela').dataTable( {
+            "language": {
+                "url": "<?= base_url(); ?>assets/js/dataTable_pt-br.json"
+            }
+        } );
+} );
+</script>
 </html>