+++ application/views/vendas/adicionarVenda.php
@@ -35,6 +35,9 @@
                                             <label for="cliente">Cliente<span class="required">*</span></label>
                                             <input id="cliente" class="span12" type="text" name="cliente" value="" />
                                             <input id="clientes_id" class="span12" type="hidden" name="clientes_id" value="" />
+                                            <div class="addclient"><?php if ($this->permission->checkPermission($this->session->userdata('permissao'), 'aCliente')) { ?>
+    <a href="<?php echo base_url(); ?>index.php/clientes/adicionar" class="btn btn-success"><i class="fas fa-plus"></i> Adicionar Cliente</a>
+<?php } ?></div>
                                         </div>
                                         <div class="span5">
                                             <label for="tecnico">Vendedor<span class="required">*</span></label>
@@ -75,11 +78,19 @@
 </div>
 <script type="text/javascript">
     $(document).ready(function() {
+        $('.addclient').hide();
         $("#cliente").autocomplete({
             source: "<?php echo base_url(); ?>index.php/vendas/autoCompleteCliente",
             minLength: 1,
+            close: function(ui) { if(ui.id == null)ui.target.value = '';},
             select: function(event, ui) {
-                $("#clientes_id").val(ui.item.id);
+                if(ui.item.label == 'Adicionar cliente...')
+                    $('.addclient').show();
+                else
+                    {
+                        $("#clientes_id").val(ui.item.id);
+                        $('.addclient').hide();
+                    }
             }
         });
         $("#tecnico").autocomplete({
@@ -128,5 +139,6 @@
         $('.editor').trumbowyg({
             lang: 'pt_br'
         });
+        $('.addclient').hide();
     });
 </script>