+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Todas as alterações serão documentadas neste arquivo
 Formato baseado em [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 e [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.26.0] - 2021-01-05
+
+## Added
+- Adiciona a cobrança ao lado do cliente;[@seitbnao](https://github.com/seitbnao)
+- Permite reenviar a cobrança por email;[@seitbnao](https://github.com/seitbnao)
+- Permite visualizar o boleto de cobrança;[@seitbnao](https://github.com/seitbnao)
+- Permite atualizar o status de cobrança;[@seitbnao](https://github.com/seitbnao)
+- Bloqueia o cliente de gerar uma cobrança, sendo restrito apenas para o administrador;[@seitbnao](https://github.com/seitbnao)
+- Faz uma alteração na tabela cobrancas, para que a mesma receba o id do cliente.[@seitbnao](https://github.com/seitbnao)
+- Altera a versão do mapos de 4.25.1 para 4.26.0
+
 ## [4.25.1] - 2021-01-04
 
 ## Fixed